backend/Recommender/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
scipy
implicit
pymongo
python-dotenv